Replace HTTPotion by HTTPoison
This commit replaces the deprecated
[HTTPotion](https://github.com/unrelentingtech/httpotion) by the hackney-based
[HTTPoison](https://github.com/edgurgel/httpoison).

Beyond the fact that we're no longer dependent on a deprecated lib, this also
enables connection pooling, which can make sequential requests to the Bot
Framework API faster.

Since this is a major change, I've also included a bit of a refactor to the
Client modules.

For the changelog
-----------------

**Replaced** `ExMicrosoftBot.Client.headers/2` by
`ExMicrosoftBot.Client.authed_headers/1`, which removes the need for fetching
the auth token to build valid request headers.

**Replaced** `ExMicrosoftBot.Client.req_opts/0` by
`ExMicrosoftBot.Client.opts/1`.

**Removed** the `http_timeout` config and introduced the `http_client_opts`
config. There doesn't seem to be an analog in HTTPoison for this as it looks
like it was used by `:ibrowse` as a default value for
[connection](https://github.com/cmullaparthi/ibrowse/blob/efd040de55f7a77d0b504e6c9de1849181bf64a6/src/ibrowse_http_client.erl#L800)
and
[request](https://github.com/cmullaparthi/ibrowse/blob/efd040de55f7a77d0b504e6c9de1849181bf64a6/src/ibrowse_http_client.erl#L920)
timeouts. Should be replaced by `http_client_opts: [timeout: X]` and
`http_client_opts: [recv_timeout: X]` respectively.

**Adds** `ExMicrosoftBot.Client.get/2`, `ExMicrosoftBot.Client.delete/2`,
`ExMicrosoftBot.Client.put/3`, and `ExMicrosoftBot.Client.post/3`.